move sleep utils to app-review
