+++ packages/@coorpacademy-app-review/src/test/index.test.tsx
@@ -5,9 +5,9 @@ import type {ExecutionContext} from 'ava';
 import React from 'react';
 import {render, fireEvent, act} from '@testing-library/react';
 import {services} from '@coorpacademy/review-services-mocks';
-import {sleep} from '@coorpacademy/review-services/es/tools/sleep';
 import type {AppOptions} from '../types/common';
 import AppReview from '..';
+import {sleep} from './utils/sleep';
 
 browserEnv({pretendToBeVisual: true});
